move mailer out of worker
